refactor(calculator): 코드의 depth를 줄이기 위한 stream 적용. 문자열 파싱 역할은 StringCalculator 클래스에서 담당하도록 수정
